+++ CHANGES.md
@@ -23,6 +23,8 @@ General:
 -   Use helm3 in inttest:registryAuth step in CI
 -   Make Magda charts helm v3 charts
 -   Add migration document for migrating Magda release v0.0.56-RC6 or eariler to v0.0.57-0
+-   Use a feature flag to turn on/off the dataset approval workflow
+-   Add internal authentication provider allow user to be authenticated locally
 
 UI:
 
@@ -50,6 +52,9 @@ UI:
 -   Use IBM Plex Sans font
 -   Fixed: error notification no longer has a white background
 -   Added `add dataset review page`
+-   Make `add dataset page` save state data to registry (unless in preview mode) rather than localStorage
+-   Change search text to say "Search for data" instead of "Search for open data"
+-   Set flag `previewAddDataset` to `false` by default
 
 Storage: